refactor: change the directory and file name for common mocks.
